% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_index_status.R
\name{check_index_status}
\alias{check_index_status}
\title{Function check the URL Index Status via the Index Metadata Endpoint}
\usage{
check_index_status(url)
}
\arguments{
\item{url}{URL you want to check
check_index_status()}
}
\description{
This function allows check the status of a given URL via the Index API Metadata Endpoint
}
\examples{
\dontrun{
check_index_status("https://my-page-in-the-index.com")
}
}
