% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_url.R
\name{index_url}
\alias{index_url}
\title{Function to send an URL to the Google Index}
\usage{
index_url(url)
}
\arguments{
\item{url}{URL you want to Index
index_url()}
}
\description{
This function allows you to send a given URL to the Google Index
}
\examples{
\dontrun{
index_url("https://my-page-in-the-index.com")
}
}
